+++ packages/react/src/components/InteractiveList/src/components/InteractiveListButtonItemSkeleton/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListButtonItemSkeleton } from "./InteractiveListButtonItemSkeleton";
+
+export { InteractiveListButtonItemSkeleton } from "./InteractiveListButtonItemSkeleton";
+export default InteractiveListButtonItemSkeleton;