+++ packages/react/src/components/InteractiveList/src/components/InteractiveListCheckboxItemSkeleton/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListCheckboxItemSkeleton } from "./InteractiveListCheckboxItemSkeleton";
+
+export { InteractiveListCheckboxItemSkeleton } from "./InteractiveListCheckboxItemSkeleton";
+export default InteractiveListCheckboxItemSkeleton;