+++ packages/react/src/components/InteractiveList/src/components/InteractiveListCheckboxItemSkeleton/interactiveListCheckboxItemSkeleton.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes } from "react";
+import { InteractiveListStructureSkeletonProperties } from "../InteractiveListStructureSkeleton/interactiveListStructureSkeleton.types";
+
+export type InteractiveListCheckboxItemSkeletonProperties =
+  InteractiveListStructureSkeletonProperties;
+
+export type InteractiveListCheckboxItemSkeletonProps =
+  InteractiveListCheckboxItemSkeletonProperties &
+    Omit<HTMLAttributes<HTMLElement>, "color">;