+++ packages/react/src/components/InteractiveList/src/components/InteractiveListRadioItemSkeleton/index.ts
@@ -1,0 +1,4 @@
+import { InteractiveListRadioItemSkeleton } from "./InteractiveListRadioItemSkeleton";
+
+export { InteractiveListRadioItemSkeleton } from "./InteractiveListRadioItemSkeleton";
+export default InteractiveListRadioItemSkeleton;