+++ packages/react/src/components/InteractiveList/src/components/InteractiveListStructureSkeleton/interactiveListStructureSkeleton.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes } from "react";
+import { InteractiveListStructureProperties } from "../InteractiveListStructure/interactiveListStructure.types";
+
+export type InteractiveListStructureSkeletonProperties =
+  InteractiveListStructureProperties;
+
+export type InteractiveListStructureSkeletonProps =
+  InteractiveListStructureSkeletonProperties &
+    Omit<HTMLAttributes<HTMLElement>, "color">;