+++ packages/react/src/components/InteractiveList/src/interactiveList.definitions.ts
@@ -1,2 +1,4 @@
 export const generateID = (name: string) =>
-  `${name.toLowerCase().replace(/ /g, "-")}_${new Crypto().randomUUID()}`;
+  `${name.toLowerCase().replace(/ /g, "-")}_${Math.floor(
+    Math.random() * Date.now() + 1
+  )}`;