+++ packages/react/src/components/InteractiveList/src/interactiveList.types.ts
@@ -2,18 +2,28 @@ import { ReactNode, HTMLAttributes } from "react";
 
 import {
   InteractiveListStructure,
+  InteractiveListStructureSkeleton,
   InteractiveListCheckboxItem,
+  InteractiveListCheckboxItemSkeleton,
   InteractiveListButtonItem,
+  InteractiveListButtonItemSkeleton,
   InteractiveListToggleItem,
   InteractiveListRadioItem,
+  InteractiveListRadioItemSkeleton,
+  InteractiveListToggleItemSkeleton,
 } from "./components";
 
 export interface InteractiveListComponents {
   Structure: typeof InteractiveListStructure;
+  StructureSkeleton: typeof InteractiveListStructureSkeleton;
   ButtonItem: typeof InteractiveListButtonItem;
+  ButtonItemSkeleton: typeof InteractiveListButtonItemSkeleton;
   CheckboxItem: typeof InteractiveListCheckboxItem;
+  CheckboxItemSkeleton: typeof InteractiveListCheckboxItemSkeleton;
   ToggleItem: typeof InteractiveListToggleItem;
+  ToggleItemSkeleton: typeof InteractiveListToggleItemSkeleton;
   RadioItem: typeof InteractiveListRadioItem;
+  RadioItemSkeleton: typeof InteractiveListRadioItemSkeleton;
 }
 
 export interface InteractiveListProperties {